add equality, fix apply
