+++ src/boot/boot.c
@@ -1,10 +1,9 @@
 /*
     A small interpreter for bootstrapping Minim.
-    Supports the most basic operations
+    Supports the most basic operations.
 */
 
 #include "boot.h"
-#include "intern.h"
 
 //
 //  Special objects
@@ -95,7 +94,7 @@ minim_object *make_closure_proc(minim_object *args, minim_object *body, minim_ob
 }
 
 //
-//  Extra constructors
+//  Extra functions
 //
 
 // Makes an association list.
@@ -117,250 +116,47 @@ minim_object *make_assoc(minim_object *xs, minim_object *ys) {
     return assoc;
 }
 
-//
-//  Forward declarations
-//
+int minim_is_eq(minim_object *a, minim_object *b) {
+    if (a == b)
+        return 1;
 
+    if (a->type != b->type)
+        return 0;
 
-//
-//  Parsing
-//
+    switch (a->type)
+    {
+    case MINIM_FIXNUM_TYPE:
+        return minim_fixnum(a) == minim_fixnum(b);
 
-int is_delimeter(int c) {
-    return isspace(c) || c == EOF || c == '(' || c == ')' || c == '"' || c == ';';
-}
-
-int is_symbol_char(int c) {
-    return isalpha(c) || c == '*' || c == '/' ||
-           c == '<' || c == '>' || c == '=' || c == '?' || c == '!';
-}
-
-int peek_char(FILE *in) {
-    int c = getc(in);
-    ungetc(c, in);
-    return c;
-}
-
-void peek_expected_delimeter(FILE *in) {
-    if (!is_delimeter(peek_char(in))) {
-        fprintf(stderr, "expected a delimeter\n");
-        exit(1);
-    }
-}
-
-void read_expected_string(FILE *in, const char *s) {
-    int c;
-
-    while (*s != '\0') {
-        c = getc(in);
-        if (c != *s) {
-            fprintf(stderr, "unexpected character: '%c'\n", c);
-            exit(1);
-        }
-        ++s;
-    }
-}
-
-void skip_whitespace(FILE *in) {
-    int c;
-
-    while ((c = getc(in)) != EOF) {
-        if (isspace(c))
-            continue;
-        else if (c == ';') {
-            // comment: ignore until newline
-            while (((c = getc(in)) != EOF) && (c != '\n'));
-        }
-        
-        // too far
-        ungetc(c, in);
-        break;
-    }
-}
-
-minim_object *read_char(FILE *in) {
-    int c;
-
-    c = getc(in);
-    switch (c) {
-    case EOF:
-        fprintf(stderr, "incomplete character literal\n");
-        break;
-    case 's':
-        if (peek_char(in) == 'p') {
-            read_expected_string(in, "pace");
-            peek_expected_delimeter(in);
-            return make_char(' ');
-        }
-        break;
-    case 'n':
-        if (peek_char(in) == 'e') {
-            read_expected_string(in, "ewline");
-            peek_expected_delimeter(in);
-            return make_char('\n');
-        }
-        break;
-    default:
-        break;
-    }
-
-    peek_expected_delimeter(in);
-    return make_char(c);
-}
-
-minim_object *read_pair(FILE *in) {
-    minim_object *car, *cdr;
-    int c;
-
-    skip_whitespace(in);
-    c = getc(in);
-    if (c == ')') {
-        // empty list
-        return minim_null;
-    }
-
-    ungetc(c, in);
-    car = read_object(in);
-
-    skip_whitespace(in);
-    c = fgetc(in);
-    if (c == '.') {
-        // improper list
-        peek_expected_delimeter(in);
-        cdr = read_object(in);
-        skip_whitespace(in);
-
-        c = getc(in);
-        if (c != ')') {
-            fprintf(stderr, "missing ')' to terminate pair");
-            exit(1);
-        }
-
-        return make_pair(car, cdr);
-    } else {
-        // list
-        ungetc(c, in);
-        cdr = read_pair(in);
-        return make_pair(car, cdr);
-    }
-}
-
-minim_object *read_object(FILE *in) {
-    char buffer[SYMBOL_MAX_LEN];
-    long num;
-    int i;
-    short sign;
-    char c;
+    case MINIM_CHAR_TYPE:
+        return minim_char(a) == minim_char(b);
     
-    skip_whitespace(in);
-    c = getc(in);
-
-    if (c == '#') {
-        // special value
-        c = getc(in);
-        switch (c) {
-        case 't':
-            return minim_true;
-        case 'f':
-            return minim_false;
-        case '\\':
-            return read_char(in);
-        }
-    } else if (isdigit(c) || ((c == '-' || c == '+') && isdigit(peek_char(in)))) {
-        // number
-        num = 0;
-        sign = 1;
-
-        // optional sign
-        if (c == '-') {
-            sign = -1;
-        } else if (c != '+') {
-            ungetc(c, in);
-        }
-
-        // magnitude
-        while (isdigit(c = getc(in))) {
-            num = (num * 10) + (c - '0');
-        }
-
-        // compose and check for delimeter
-        num *= sign;
-        if (!is_delimeter(c)) {
-            fprintf(stderr, "expected a delimeter\n");
-            exit(1);
-        }
-
-        ungetc(c, in);
-        return make_fixnum(num);
-    } else if (is_symbol_char(c) || ((c == '+' || c == '-') && is_delimeter(peek_char(in)))) {
-        // symbol
-        i = 0;
-
-        while (is_symbol_char(c) || isdigit(c) || c == '+' || c == '-') {
-            if (i < SYMBOL_MAX_LEN - 1) {
-                buffer[i++] = c;
-            } else {
-                fprintf(stderr, "symbol is too long, max allowed %d characters", SYMBOL_MAX_LEN);
-                exit(1);
-            }
-            c = getc(in);
-        }
-
-        if (!is_delimeter(c)) {
-            fprintf(stderr, "expected a delimeter\n");
-            exit(1);
-        }
-
-        ungetc(c, in);
-        buffer[i] = '\0';
-        return intern_symbol(symbols, buffer);
-    } else if (c == '"') {
-        // string
-        i = 0;
-        while ((c = getc(in)) != '"') {
-            if (c == '\\') {
-                c = getc(in);
-                if (c == 'n') {
-                    c = '\n';
-                } else if (c == 't') {
-                    c = '\t';
-                } else if (c == '\\') {
-                    c = '\\';
-                } else {
-                    fprintf(stderr, "unknown escape character: %c\n", c);
-                    exit(1);
-                }
-            } else if (c == EOF) {
-                fprintf(stderr, "non-terminated string literal\n");
-                exit(1);
-            }
-
-            if (i < SYMBOL_MAX_LEN - 1) {
-                buffer[i++] = c;
-            } else {
-                fprintf(stderr, "string is too long, max allowed %d characters", SYMBOL_MAX_LEN);
-                exit(1);
-            }
-        }
-
-        buffer[i] = '\0';
-        return make_string(buffer);
-    } else if (c == '(') {
-        // empty list or pair
-        return read_pair(in);
-    } else if (c == '\'') {
-        // quoted expression
-        return make_pair(quote_symbol, make_pair(read_object(in), minim_null));
-    } else if (c == EOF) {
-        return NULL;
-    } else {
-        fprintf(stderr, "unexpected input: %c\n", c);
-        exit(1);
+    default:
+        return 0;
     }
-
-    fprintf(stderr, "unreachable");
-    exit(1);
 }
+
+int minim_is_equal(minim_object *a, minim_object *b) {
+    if (minim_is_eq(a, b))
+        return 1;
+
+    if (a->type != b->type)
+        return 0;
+
+    switch (a->type)
+    {
+    case MINIM_STRING_TYPE:
+        return strcmp(minim_string(a), minim_string(b)) == 0;
+
+    case MINIM_PAIR_TYPE:
+        return minim_is_equal(minim_car(a), minim_car(b)) &&
+               minim_is_equal(minim_cdr(a), minim_cdr(b));
+    
+    default:
+        return 0;
+    }
+}
+
 
 //
 //  Environments
@@ -503,6 +299,18 @@ minim_object *string_to_symbol_proc(minim_object *args) {
     return intern_symbol(symbols, minim_string(minim_car(args)));
 }
 
+minim_object *eq_proc(minim_object *args) {
+    minim_object *a = minim_car(args);
+    minim_object *b = minim_cadr(args);
+    return minim_is_eq(a, b) ? minim_true : minim_false;
+}
+
+minim_object *equal_proc(minim_object *args) {
+    minim_object *a = minim_car(args);
+    minim_object *b = minim_cadr(args);
+    return minim_is_equal(a, b) ? minim_true : minim_false;
+}
+
 minim_object *eval_proc(minim_object *args) {
     fprintf(stderr, "eval: should not be called directly");
     exit(1);
@@ -585,6 +393,16 @@ static minim_object *let_vals(minim_object *bindings) {
             make_pair(minim_car(minim_cdar(bindings)), let_vals(minim_cdr(bindings))));
 }
 
+static minim_object *apply_args(minim_object *args) {
+    if (minim_is_null(args)) {
+        return minim_null;
+    } else if (minim_is_null(minim_cdr(args))) {
+        return minim_car(args);
+    } else {
+        return make_pair(minim_car(args), apply_args(minim_cdr(args)));
+    }
+}
+
 static minim_object *eval_exprs(minim_object *exprs, minim_object *env) {
     if (minim_is_null(exprs)) {
         return minim_null;
@@ -620,7 +438,7 @@ loop:
         return minim_void;
     } else if (is_let(expr)) {
         env = extend_env(let_vars(minim_cadr(expr)), let_vals(minim_cadr(expr)), env);
-        expr = minim_car(minim_cddr(expr));
+        expr = make_pair(begin_symbol, minim_car(minim_cddr(expr)));
         goto loop;
     } else if (is_if(expr)) {
         if (minim_is_true(eval_expr(minim_cadr(expr), env)))
@@ -695,7 +513,7 @@ application:
             // special case for `apply`
             if (minim_prim_proc(proc) == apply_proc) {
                 proc = minim_car(args);
-                args = minim_cdr(args);
+                args = apply_args(minim_cdr(args));
                 goto application;
             }
 
@@ -720,104 +538,6 @@ application:
 }
 
 //
-//  Printing
-//
-
-void write_pair(FILE *out, minim_pair_object *p) {
-    write_object(out, minim_car(p));
-    if (minim_is_pair(minim_cdr(p))) {
-        fputc(' ', out);
-        write_pair(out, ((minim_pair_object *) minim_cdr(p)));
-    } else if (minim_is_null(minim_cdr(p))) {
-        return;
-    } else {
-        fprintf(out, " . ");
-        write_object(out, minim_cdr(p));
-    }
-}
-
-void write_object(FILE *out, minim_object *o) {
-    switch (o->type) {
-    case MINIM_NULL_TYPE:
-        fprintf(out, "()");
-        break;
-    case MINIM_TRUE_TYPE:
-        fprintf(out, "#t");
-        break;
-    case MINIM_FALSE_TYPE:
-        fprintf(out, "#f");
-        break;
-    case MINIM_EOF_TYPE:
-        fprintf(out, "#<eof>");
-        break;
-    case MINIM_VOID_TYPE:
-        fprintf(out, "#<void>");
-        break;
-
-    case MINIM_SYMBOL_TYPE:
-        fprintf(out, "%s", minim_symbol(o));
-        break;
-    case MINIM_FIXNUM_TYPE:
-        fprintf(out, "%ld", minim_fixnum(o));
-        break;
-    case MINIM_CHAR_TYPE:
-        int c = minim_char(o);
-        switch (c) {
-        case '\n':
-            fprintf(out, "#\\newline");
-            break;
-        case ' ':
-            fprintf(out, "#\\space");
-            break;
-        default:
-            fprintf(out, "#\\%c", c);
-        }
-        break;
-    case MINIM_STRING_TYPE:
-        char *str = minim_string(o);
-        fputc('"', out);
-        while (*str != '\0') {
-            switch (*str) {
-            case '\n':
-                fprintf(out, "\\n");
-                break;
-            case '\t':
-                fprintf(out, "\\t");
-                break;
-            case '\\':
-                fprintf(out, "\\\\");
-                break;
-            default:
-                fputc(*str, out);
-                break;
-            }
-            ++str;
-        }
-        fputc('"', out);
-        break;
-    case MINIM_PAIR_TYPE:
-        fputc('(', out);
-        write_pair(out, ((minim_pair_object *) o));
-        fputc(')', out);
-        break;
-    case MINIM_PRIM_PROC_TYPE:
-    case MINIM_CLOSURE_PROC_TYPE:
-        fprintf(out, "#<procedure>");
-        break;
-    case MINIM_OUTPUT_PORT_TYPE:
-        fprintf(out, "#<output-port>");
-        break;
-    case MINIM_INPUT_PORT_TYPE:
-        fprintf(out, "#<input-port>");
-        break;
-    
-    default:
-        fprintf(stderr, "cannot write unknown object");
-        exit(1);
-    }
-}
-
-//
 //  Interpreter initialization
 //
 
@@ -839,6 +559,9 @@ void populate_env(minim_object *env) {
     add_procedure("string->number", string_to_number_proc);
     add_procedure("symbol->string", symbol_to_string_proc);
     add_procedure("string->symbol", string_to_symbol_proc);
+
+    add_procedure("eq?", eq_proc);
+    add_procedure("equal?", equal_proc);
 
     add_procedure("interaction-environment", interaction_environment_proc);
     add_procedure("null-environment", empty_environment_proc);
