+++ src/boot/boot.h
@@ -7,6 +7,7 @@
 #define _BOOT_H_
 
 #include "../minim.h"
+#include "intern.h"
 
 #define MINIM_BOOT_VERSION      "0.1.0"
 
@@ -19,10 +20,11 @@
 void minim_boot_init();
 minim_object *make_env();
 minim_object *eval_expr(minim_object *expr, minim_object *env);
+
 minim_object *read_object(FILE *in);
 void write_object(FILE *out, minim_object *o);
 
 extern minim_object *global_env;
-
+extern intern_table *symbols;
 
 #endif  // _BOOT_H_
